requirements.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
liger-kernel
numpy
pandas
peft
pyarrow>=15.0.0
pybind11
pylatexenc
pylint==3.3.6
ray[default]
tensordict<=0.6.2
torchdata
transformers
vllm==0.8.2
wandb
